requirements.txt
python-telegram-bot
python-telegram-bot[job-queue]
nest_asyncio
beautifulsoup4
requests